Player input moved to game module
