+++ src/entity/entity.c
@@ -1,0 +1,1 @@
+#include "entity.h"
