+++ src/views/game/game.h
@@ -1,7 +1,9 @@
-
+#include "controller.h"
+#include "entity/entity.h"
 
 #ifndef GAME_H
 #define GAME_H
 
+void update_player_input(Input *, Entity *);
 
 #endif
