+++ src/views/pause/pause.h
@@ -1,0 +1,5 @@
+#ifndef PAUSE_H
+#define PAUSE_H
+
+
+#endif
